fix: Changed how the grid templates work
Now, if you provide only the initial template and a tablet template, the mobile template will default to the tablet's allowing you to just set the tabletTemplate if you just want one UI break point.
Moreover, another build was run to test the latest changes and save everything for in one commit, and changed the OBC examples to work just fine on tablet screens.